feat(EmailTemplateDetailView): add view
