+++ src/sa_api_v2/views/__init__.py
@@ -1,2 +1,3 @@
 from .base_views import *
+from .email_templates import *
 from .bulk_data_views import *
